+++ example/lib/lpc/lpc_game.dart
@@ -29,7 +29,7 @@ class LPCGame extends StatelessWidget {
         keyboardConfig: KeyboardConfig(),
         directional: JoystickDirectional(),
       ),
-      map: MapWordByTiled(
+      map: WorldMapByTiled(
         'tiled/mapa2.json',
         forceTileSize: Vector2(32, 32),
       ),